fps controls framework done
Made a fps controls framework

Allows for fps camera movement, and also player movement. A fairly simple implementation, but a good framework to build upon later

For now, to unpause the game, just click on the screen, and either esc or tab out to pause the game. Game will freeze when paused, but will resume movement when unpaused (done intentionally because that's how pausing works. see isPaused). Made use of cursor locking.